+++ src/display/autotiles.cpp
@@ -1,12 +1,15 @@
 struct StaticRect { float x, y, w, h; };
+
 /*
 Calculated coordinates for autotile pieces. 
 Autotiles are 3 x 4 tiles in size, and they're sliced into
 half pieces and rearranged to form autotiles. This
 pre-calculates all of the rects for these slices.
 */
-static const float tileSize = 16;
-static const float halfTile = tileSize / 2;
+// static const float tileSize = 16;
+// static const float halfTile = tileSize / 2;
+static const int tileSize = 16;
+static const int halfTile = tileSize / 2;
 
 extern const StaticRect autotileRects[] =
 {
